+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Fixes a bug in `VcsExtent`, where feature were not reset, if extent creation was canceled
 - Fixes a bug where selecting a HEX color would fail and fallback to black
 - Fixes a bug where selecting an EPSG in PositionDisplay could fail
+- Fixes a bug, where `Select all` was disabled after adding the first entry in a category.
 
 # 6.0.4